refactor: rebuild session on a per-network backend abstraction with reactive state
Extract a NetworkBackend/AccountBackend layer (GDK, Greenlight and LWK
implementations, plus an LWK Waterfalls->Esplora->Electrum fallback client) and
move per-network and per-account state out of GdkSession into the backends as
StateFlows. GdkSession becomes an orchestrator: session-level accessors derive
from the per-backend flows, single-network lookups route through fail-fast
reified backend lookups, and all backends connect in parallel with per-network
failures recorded in failedNetworksStateFlow. Backend teardown is a suspend
disconnect (AutoCloseable/close dropped) that clears accounts and backend maps.

Replace the imperative balance push (updateAccountsAndBalances, walletAssets)
with reactive sources: account backends cache their assets inside getBalance,
and consumers read through a use-case layer built on a shared
DataStateObservableUseCase base (mutex-serialized work on Dispatchers.Default,
DataState results, shared observe()):

- GetWalletAssetsUseCase aggregates per-account balances
- GetWalletTransactionsUseCase / GetAccountTransactionsUseCase aggregate
  paginated per-account pages, refresh on debounced NetworkEvents, and derive
  hasMore from per-account page fullness
- GetReceiveAddressUseCase resolves receive addresses per backend

Remove the active-account concept and the Account Overview, WatchOnly and
WalletAssets screens; the wallet overview lists accounts directly. Split
Address into an interface with GdkAddress and GlAddress implementations.

Fix the ANR and event-storm class of bugs along the way: GDK JNI calls run off
the main thread via gdkContext (no more runBlocking in denomination/exchange
paths), Denomination and unit/fiat accessors read the cached settings
StateFlow, duplicate Lightning NewBlock events are deduped with
distinctUntilChanged, networkEventsFlow no longer replays (nor drops the first
live event), confirmed transactions are not refetched on every block, and the
overview skips accounts whose balance fails to resolve instead of crashing.

+++ data/src/commonMain/kotlin/com/blockstream/data/gdk/data/SubaccountEvent.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class SubaccountEvent(
     @SerialName("pointer")
-    val pointer: Int = 0,
+    val pointer: Long = 0,
     @SerialName("event_type")
     val eventType: String = "",
 ) {
